+++ packages/vendure-plugin-primary-collection/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.0.0 (2024-06-21)
 
 - Make `primaryCollection` custom field channel aware.
+- Resolve `product.breadcrumbs` when a product has a primary collection set
 
 # 1.5.0 (2024-06-21)